Fix: quizSettings component 삭제
