+++ src/pages/quiz-settings/index.tsx
@@ -1,10 +1,15 @@
 import { useForm } from "react-hook-form";
 import useAuthStore from "@/store/AuthStore";
-import QuizSettingsForm, {
-  QuizSettingsValues,
-} from "@/components/QuizSettingForm";
 import { useEffect } from "react";
 import { useRouter } from "next/router";
+import FormInput from "@/components/FormInput";
+import FilledButton from "@/components/ui/Button/FilledButton";
+
+
+export interface QuizSettingsValues {
+  securityAnswer: string;
+  securityQuestion: string;
+}
 
 const QuizSettings = () => {
   const {
@@ -15,7 +20,7 @@ const QuizSettings = () => {
     mode: "onSubmit",
   });
 
-  const { createProfile } = useAuthStore();
+  const { createProfile,isLoggedIn,user } = useAuthStore();
   const router = useRouter();
 
   const handleProfileCreation = async (data: QuizSettingsValues) => {
@@ -24,22 +29,67 @@ const QuizSettings = () => {
     router.push(`/wiki/${user?.profile?.name}`);
   };
 
-  useEffect(()=> {
-   const {isLoggedIn, user} = useAuthStore.getState()
-   if (isLoggedIn && user?.profile) {
-    router.push(`/wiki/${user.profile.name}`)
-   } else {
-    router.push('/login')
-   }
-  },[])
+  useEffect(() => {
+
+    if (!isLoggedIn) {
+      router.push("/login");
+    } else if (user?.profile) {
+      router.push(`/wiki/${user.profile.name}`);
+    }
+  }, [isLoggedIn, user, router]);
 
   return (
-    <QuizSettingsForm
-      onSubmit={handleSubmit(handleProfileCreation)}
-      register={register}
-      errors={errors}
-      isSubmitting={isSubmitting}
-    />
+    <div className="flex items-center justify-center min-h-screen Mobile:px-5 bg-background">
+      <div className="w-full max-w-md">
+        <h2 className="text-2xl font-semibold mb-10">
+          좋아요!
+          <br />
+          이제 위키를 만들어 볼까요?
+        </h2>
+        <div className="flex flex-col">
+          <div className="py-6 relative border-t-2">
+            <h3 className="text-xl font-semibold">
+              원하는 사람만 편집할 수 있게 하기
+            </h3>
+            <span className="text-md text-gray-400">
+              친구들만 대답할 수 있는 질문을 등록해 보세요.
+            </span>
+          </div>
+        </div>
+        <form
+          onSubmit={handleSubmit(handleProfileCreation)}
+          className="flex flex-col gap-6"
+        >
+          <FormInput
+            id="securityQuestion"
+            placeholder="질문을 입력해 주세요"
+            type="text"
+            register={register("securityQuestion", {
+              required: "질문을 입력해 주세요",
+            })}
+            error={errors.securityQuestion}
+          />
+          <FormInput
+            id="securityAnswer"
+            placeholder="답을 입력해주세요"
+            type="text"
+            register={register("securityAnswer", {
+              required: "답을 입력해주세요",
+            })}
+            error={errors.securityAnswer}
+          />
+          <div className="border-t-2 py-6">
+            <FilledButton
+              fullWidth={true}
+              type="submit"
+              disabled={isSubmitting}
+            >
+              위키 만들기
+            </FilledButton>
+          </div>
+        </form>
+      </div>
+    </div>
   );
 };
